Dirty workaround for importStep error
Currently, if you add profile-Products.CMFNotification:default as a GS
dependency and configure it via a cmfnotification.xml file it will fail
because it tries to configure the portal_notification tool before
installing it. This change hides that problem by not throwing an error
if the portal_notification tool isn't installed. It is still possible to
use the quickinstaller to install it and then run the import steps to
configure it:

    qi = api.portal.get_tool('portal_quickinstaller')
    qi.installProduct('Products.CMFNotification')
    context.runImportStepFromProfile(
        default_profile,
        'cmfnotification',
        run_dependencies=False,
    )